Fix issue with percentage sliders
Using the InputNumber of NumberSlider caused a bug
where only the first digit of the input would be saved.

This was due to Form.Item intercepting the event
and overriding the `onChange` handler of NumberSlider.

Removing the `name` prop from Form.Item fixes the bug
by delegating this responsibility to NumberSlider's `onChange`
event.

Accessibility is unaffected by this change.