Solid foundation for React testing library usage
Going to get started with testing the EthAddress field. Needed
to setup a test Antd form in a custom testable component. A
little less than ideal, since we should strive to test REAL forms,
but it's not a big deal.

Getting stuck with mocking the readProvider, but should be resolvable.
Ran into other issues with Jest and lingui that should be resolved now.